# [MPSInductor] Speedup `sum`/`prod` reductions