--- a/c10/metal/reduction_utils.h
+++ b/c10/metal/reduction_utils.h
@@ -6,27 +6,52 @@
 namespace c10 {
 namespace metal {
 
+constant constexpr unsigned simdgroup_size = 32;
+
 template <typename T>
-opmath_t<T> threadgroup_sum(threadgroup T* data, unsigned size) {
-  // TODO: This should be moved to the callee
-  ::metal::threadgroup_barrier(::metal::mem_flags::mem_threadgroup);
-  opmath_t<T> rc = data[0];
-  // TODO: Use `simd_shuffle_down`
-  for (unsigned idx = 1; idx < size; ++idx) {
-    rc += data[idx];
+opmath_t<T> threadgroup_sum(
+    threadgroup T* data,
+    T val,
+    unsigned idx,
+    unsigned size) {
+  auto rc = ::metal::simd_sum(val);
+  if (idx % simdgroup_size == 0) {
+    data[idx / simdgroup_size] = rc;
   }
-  return rc;
+  if (size > simdgroup_size) {
+    ::metal::threadgroup_barrier(::metal::mem_flags::mem_threadgroup);
+    if (idx < ((size + simdgroup_size - 1) / simdgroup_size)) {
+      auto rc1 = ::metal::simd_sum(data[idx]);
+      if (idx == 0) {
+        data[0] = rc1;
+      }
+    }
+  }
+  ::metal::threadgroup_barrier(::metal::mem_flags::mem_threadgroup);
+  return data[0];
 }
 
 template <typename T>
-opmath_t<T> threadgroup_prod(threadgroup T* data, unsigned size) {
-  // TODO: This should be moved to the callee
-  ::metal::threadgroup_barrier(::metal::mem_flags::mem_threadgroup);
-  opmath_t<T> rc = data[0];
-  for (unsigned idx = 1; idx < size; ++idx) {
-    rc *= data[idx];
+opmath_t<T> threadgroup_prod(
+    threadgroup T* data,
+    T val,
+    unsigned idx,
+    unsigned size) {
+  auto rc = ::metal::simd_product(val);
+  if (idx % simdgroup_size == 0) {
+    data[idx / simdgroup_size] = rc;
   }
-  return rc;
+  if (size > simdgroup_size) {
+    ::metal::threadgroup_barrier(::metal::mem_flags::mem_threadgroup);
+    if (idx < ((size + simdgroup_size - 1) / simdgroup_size)) {
+      auto rc1 = ::metal::simd_product(data[idx]);
+      if (idx == 0) {
+        data[0] = rc1;
+      }
+    }
+  }
+  ::metal::threadgroup_barrier(::metal::mem_flags::mem_threadgroup);
+  return data[0];
 }
 
 template <typename T>